backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Database
chromadb==0.4.22

# AI & NLP
langchain==0.1.6
langchain-community==0.0.20
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.2
pandas==2.1.4
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10

# Internet Search
duckduckgo-search==4.1.1
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

